+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and to the [CHANGELOG recommendations](http://keepachangelog.com/).
 ### [0.2.0] - (2020-11-20)
 - Merge ressources offered by different sources
 - Ordered output
+- Linked to OpenJustice document api
 
 ### [0.1.1] - (2020-11-11)
 - Added changelog
